fix: rename keyboard layouts
